backend/requirement.txt
arabic-reshaper==3.0.0
asgiref==3.6.0
asn1crypto==1.5.1
certifi==2023.5.7
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.1.0
click==8.1.7
cryptography==42.0.0
cssselect2==0.7.0
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.6
django-cors-headers==3.14.0
django-import-export==3.2.0
django-resized==1.0.2
djangorestframework==3.14.0
et-xmlfile==1.1.0
html5lib==1.1
idna==3.4
lxml==5.1.0
MarkupPy==1.14
numpy==1.26.3
odfpy==1.4.1
openpyxl==3.1.2
oscrypto==1.3.0
pandas==2.1.0
Pillow==9.4.0
pycparser==2.21
pyHanko==0.21.0
pyhanko-certvalidator==0.26.3
PyJWT==1.7.1
pypdf==4.0.0
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-slugify==8.0.1
pytz==2022.7.1
PyYAML==6.0.1
qrcode==7.4.2
reportlab==4.0.9
requests==2.31.0
six==1.16.0
sqlparse==0.4.3
svglib==1.5.1
tablib==3.5.0
text-unidecode==1.3
tinycss2==1.2.1
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
uritools==4.0.2
urllib3==2.0.3
webencodings==0.5.1
xhtml2pdf==0.2.9
xlrd==2.0.1
xlwt==1.3.0
---------------------------------------------------------------------------------------------
- Gestion Compte Par Compagne
- Principe de Gestion Password Nomenclature taille et Caractères
- Table Gestion Table Standard pour la validation des Données


add Logo COH
type projet Agro et PSE

Remplacer Carte par Geoportail

reduire a deux Chiffre le tx de CO2

Connaissement Numéro de Connaissement :

origine de l'arbre pour la description des Arbres Existants

especes plantées avec e muet
liste espece et non epèces

Remplacer Suivis par planting
certiif COH


remplacer AGR-OMAP par AGROMAP sur le Fichier PDF

---------------------------------------------------------------------------------------------------

Echange avec Mr Weil commercial chez Google :

 - outils de gestion avec les Pack


unité Production (t)

Nombre de Parcelle(s)

Nombre de Sections

status foncier :

Indice de couleur pour les parcelles dont la superficie est Supérieur a 4 ha

au dela de l'aspect carbone contribuer a la biodiversité et faune également


Meet with Google
 -

Gestion Formulaire enquêtes Sociales :

RDUE COOP

Evaluateur / Enquéteur
Entité / Organisation
Produits / Produits Commercialisés (Liste déroulante produits liés au RDUE)
Chaine D'appro (source reseau de producteur ou producteur externe a la coopérative)
pays d'origine (liste de pays déroulante pays frontaliers)
zones a risque (Orpaiilage, deforesattion, travail des enfants, exploitation illegale de bois, aucune) - zone d'activités portant a contreverse

Mesures Correctives (Liste déroulante)


Enquete social
catégoriser les Enquetes

Info sur le Menage
Info sur la parcelles / Exploitation


Infos Menage :
Nbre Epouse
Tranche d'age enfant
Nbre enfant
Moyen paiement (Mobile money/Compte Bancaire)  - si Mobile Money Preciser le numero money

Info sur la parcelles / Exploitation:
 - Utilisation d'intrant (Produit Phyto Sanitaire)
 - Nombre de Main d'Oeuvre (Nbre femme/ nb homme)- Tranche d'age
 - type de Main d'Oeuvre (Permanant/occasionnel) - Homme / femme
 - Salaire Moyen main d'Oeuvre - Mode de Paiement main D'Oeuvre - Salaire/Par Production

 - Production Concernées
 - Type de Production - Production secondaire

Intrants
Produit phyto (Herbicide, fongicite Insecticides)
Angrais  (Mineral, Organique)



 Intrant Bio
  engrais organique
  engrais bio
  semences Biologiques




Document de Propriete de la PARCELLE
Aucun
Acte de Vente
Certificat Foncier
Attestation Villageoise
ACD

Preciser le Document

STANDARD DE CERTIFICATION
RA
UTZ
ETC
RSPO

Donner la possibilite de choisir plusieurs Certificat

ne pas rendre la superficie Obligatoire

------------------------------------------------------------------------------------------------------------------------
Rencontre avec My EasyFarm

resolution satellitaire sentinelle 2 avec une résolution 10m/10m

 - Importation données Groupée
 - Language de Developpement
 - Est-ce que MyEasyFarm utilisent des Application ou API tierce developpées developer par d'autre Structure ? si comment sont gérer le principe de Confidentialité entre un partenaire et Easyfarm
 - est-ce qu'il y a une posibilite de Décentraliser le principe d'hébergement de l'application si oui comment procéder ?


 -----------------------------------------------------------------------------------------------------------------------

 Proposez un rapport pour la Coopérative

Extraction Infos Anonymiser
Extraction Infos Complête
Extraction infos par case a Cocher


password Coopaahs: coopaahs2024, EMail : anitchecoop@gmail.com
password Agrial : Agri@l2024#, EMail : agrialcopca@gmail.com





// Declare map using EPSG3857 projection (default is also 3857, so just optionnal here) and set center
const center = [38, 20.472157];
const map = L.map('map', {
    crs: L.CRS.EPSG3857
}).setView(center, 4);

// Define wmsOptions for wmsLayer
const wmsOptions = {
    layers: 'Landsat-8',
    transparent: true,
    format: 'image/png'
}

// WMS Layer constructor
const wmsLayer = L.tileLayer.wms('https://kade.si/cgi-bin/mapserv', wmsOptions);

// add To the map
wmsLayer.addTo(map);



https://github.com/Igor-Vladyka/leaflet.browser.print

